changed ffmpeg version to ffmpeg-3.1.4

+++ libavcodec/h264dsp.h
@@ -28,12 +28,11 @@
 #define AVCODEC_H264DSP_H
 
 #include <stdint.h>
-#include <stddef.h>
 
-typedef void (*h264_weight_func)(uint8_t *block, ptrdiff_t stride, int height,
+typedef void (*h264_weight_func)(uint8_t *block, int stride, int height,
                                  int log2_denom, int weight, int offset);
 typedef void (*h264_biweight_func)(uint8_t *dst, uint8_t *src,
-                                   ptrdiff_t stride, int height, int log2_denom,
+                                   int stride, int height, int log2_denom,
                                    int weightd, int weights, int offset);
 
 /**
